What is done
Fix double tap issue on Mobile
Currently, you must to double tab a video and is very annoying for the users.

Fetch YouTube API script on demand
The client fetch the YouTube API Library only when the user click a placeholder video by the first time.

It doesn't use iframes directly
iframes are using internally through the YouTube API Library

Keep the performance stunning
It doesn't impact to the perfomance because the library is not being fetched at the same time that the page is being loaded